py/requirements.txt
requests>=2.28.0
PyYAML>=6.0
mcp[cli]
beautifulsoup4>=4.11.0
httpx>=0.23.0